+++ src/mcu_stm32h745/CM7/boot/bootloader_common.h
@@ -60,11 +60,11 @@ extern IWDG_HandleTypeDef hiwdg;
 // rule applies here verbatim.
 #define FIRMWARE_VERSION_MAJOR 0
 #define FIRMWARE_VERSION_MINOR 1
-#define FIRMWARE_VERSION_PATCH 1
+#define FIRMWARE_VERSION_PATCH 2
 
 #define BOOTLOADER_VERSION_MAJOR 0
 #define BOOTLOADER_VERSION_MINOR 1
-#define BOOTLOADER_VERSION_PATCH 1
+#define BOOTLOADER_VERSION_PATCH 2
 
 extern const uint8_t HMAC_KEY[32];